fix: master footer file created
